requirements.txt
# selenium==4.1.5
# webdriver-manager==3.6.1
requests==2.27.1
python-telegram-bot==13.11